United TransactionChains and SystemCounters classes into the TransactionScheduler class. Need to implement RETRY Chains mechanism.

+++ GPSS.NET/GPSS.NET/StandardAttributes/ISystemAttributes.cs
@@ -12,7 +12,7 @@ namespace GPSS.StandardAttributes
         /// <remarks>
         /// GPSS World TG1 SNA.
         /// </remarks>
-        double TerminationCount { get; }
+        int TerminationCount { get; }
 
         /// <summary>
         /// Value of absolute system clock. Simulated time since last CLEAR.
